feat(MeshService): add grace period before deleting generated MeshServices on universal (#11018)

+++ pkg/core/resources/apis/meshservice/generate/generator.go
@@ -2,6 +2,7 @@ package generate
 
 import (
 	"context"
+	"maps"
 	"reflect"
 	"slices"
 	"strings"
@@ -25,24 +26,27 @@ import (
 )
 
 const (
-	managedByValue string = "meshservice-generator"
+	managedByValue                  string = "meshservice-generator"
+	deletionGracePeriodStartedLabel string = "kuma.io/deletion-grace-period-started-at"
 )
 
 // Generator generates MeshService objects from Dataplane resources created on
 // universal.
 type Generator struct {
-	logger     logr.Logger
-	interval   time.Duration
-	metric     prometheus.Summary
-	resManager manager.ResourceManager
-	meshCache  *mesh.Cache
+	logger              logr.Logger
+	generateInterval    time.Duration
+	deletionGracePeriod time.Duration
+	metric              prometheus.Summary
+	resManager          manager.ResourceManager
+	meshCache           *mesh.Cache
 }
 
 var _ component.Component = &Generator{}
 
 func New(
 	logger logr.Logger,
-	interval time.Duration,
+	generateInterval time.Duration,
+	deletionGracePeriod time.Duration,
 	metrics core_metrics.Metrics,
 	resManager manager.ResourceManager,
 	meshCache *mesh.Cache,
@@ -56,11 +60,12 @@ func New(
 		return nil, err
 	}
 	return &Generator{
-		logger:     logger,
-		interval:   interval,
-		metric:     metric,
-		resManager: resManager,
-		meshCache:  meshCache,
+		logger:              logger,
+		generateInterval:    generateInterval,
+		deletionGracePeriod: deletionGracePeriod,
+		metric:              metric,
+		resManager:          resManager,
+		meshCache:           meshCache,
 	}, nil
 }
 
@@ -175,22 +180,60 @@ func (g *Generator) generate(ctx context.Context, mesh string, dataplanes []*cor
 			log.Info("Port conflict for a kuma.io/service tag, ports must be identical across Dataplane inbounds for a given kuma.io/service", "dps", dps)
 		}
 		delete(meshservicesByName, meshService.GetMeta().GetName())
-		if newMeshService != nil && servicesDiffer(meshService.Spec, newMeshService) {
+		gracePeriodStartedAtText, hasGracePeriodLabel := meshService.GetMeta().GetLabels()[deletionGracePeriodStartedLabel]
+
+		servicesDiffer := newMeshService != nil && servicesDiffer(meshService.Spec, newMeshService)
+		if newMeshService != nil && (servicesDiffer || hasGracePeriodLabel) {
 			meta := meshService.GetMeta()
 			meshService = meshservice_api.NewMeshServiceResource()
 			meshService.Meta = meta
 			meshService.Spec = newMeshService
-			if err := g.resManager.Update(ctx, meshService); err != nil {
+
+			// Unset the grace period by deleting the label
+			newLabels := maps.Clone(meshService.GetMeta().GetLabels())
+			delete(newLabels, deletionGracePeriodStartedLabel)
+
+			if err := g.resManager.Update(ctx, meshService, store.UpdateWithLabels(newLabels)); err != nil {
 				log.Error(err, "couldn't update MeshService")
 				continue
 			}
-			log.Info("updated MeshService")
-		} else if newMeshService == nil {
-			if err := g.resManager.Delete(ctx, meshservice_api.NewMeshServiceResource(), store.DeleteBy(model.MetaToResourceKey(meshService.GetMeta()))); err != nil {
-				log.Error(err, "couldn't delete MeshService")
-				continue
+			var reasons []string
+			if servicesDiffer {
+				reasons = append(reasons, "spec changed")
 			}
-			log.Info("deleted MeshService")
+			if hasGracePeriodLabel {
+				reasons = append(reasons, "no longer scheduled for deletion")
+			}
+			log.Info("updated MeshService", "reasons", reasons)
+		} else if newMeshService == nil {
+			gracePeriodStartedAt, err := time.Parse(time.RFC3339, gracePeriodStartedAtText)
+			if hasGracePeriodLabel && err == nil {
+				// If we have a valid grace period set, check if it's expired
+				if time.Since(gracePeriodStartedAt) > g.deletionGracePeriod {
+					if err := g.resManager.Delete(ctx, meshservice_api.NewMeshServiceResource(), store.DeleteBy(model.MetaToResourceKey(meshService.GetMeta()))); err != nil {
+						log.Error(err, "couldn't delete MeshService")
+						continue
+					}
+					log.Info("deleted MeshService")
+				}
+			} else {
+				// Start the grace period if we don't have the label or it's invalid
+				if hasGracePeriodLabel && err != nil {
+					log.Info("couldn't parse grace period label, ignoring", "value", gracePeriodStartedAtText)
+				}
+				nowText, err := time.Now().MarshalText()
+				if err != nil {
+					log.Error(err, "couldn't marshal time.Now as text, this shouldn't be possible")
+					continue
+				}
+				newLabels := maps.Clone(meshService.GetMeta().GetLabels())
+				newLabels[deletionGracePeriodStartedLabel] = string(nowText)
+				if err := g.resManager.Update(ctx, meshService, store.UpdateWithLabels(newLabels)); err != nil {
+					log.Error(err, "couldn't update MeshService")
+					continue
+				}
+				log.Info("MeshService deletion grace period started", "period", g.deletionGracePeriod.String())
+			}
 		}
 	}
 	for name, meshServices := range meshservicesByName {
@@ -222,7 +265,7 @@ func (g *Generator) NeedLeaderElection() bool {
 
 func (g *Generator) Start(stop <-chan struct{}) error {
 	g.logger.Info("starting")
-	ticker := time.NewTicker(g.interval)
+	ticker := time.NewTicker(g.generateInterval)
 	ctx := user.Ctx(context.Background(), user.ControlPlane)
 
 	for {
